break down files
