+++ data_privacy_law/db_manager/data/bill_info.csv
@@ -1,0 +1,9 @@
+Title,Date,Type,Sector,State,Path,Filename
+Texas: Act prohibiting social media use by children,09012025,State level sectoral,Children’s Data Protection,Texas,./pdfs/Texas/HB 186 Social_media_children.pdf,HB 186 Social_media_children.pdf
+Texas: Prohibition on Menstrual Health Data Search Warrants,09012025,State level sectoral,Health,Texas,./pdfs/Texas/HB 1426 Mentrual Health Info.pdf,HB 1426 Mentrual Health Info.pdf
+Texas: Act relating to maternal mortality information,09012025,Comprehensive State level,Health,Texas,./pdfs/Texas/HB 1989 Maternal Mortality.pdf,HB 1989 Maternal Mortality.pdf
+Texas: Act relating to patient medical records and discharge duties,09012025,State level sectoral,Health,Texas,./pdfs/Texas/HB 2264 Medical Records.pdf,HB 2264 Medical Records.pdf
+Texas: Act relating to reproductive health information collection and civil penalties,09012025,State level sectoral,Health,Texas,./pdfs/Texas/HB 2831 Reproductive health info.pdf,HB 2831 Reproductive health info.pdf
+Texas: Act relating to smart device data collection transparency,09012025,State level sectoral,Telecommunications & Technology,Texas,./pdfs/Texas/SB 726 Smart Devices.pdf,SB 726 Smart Devices.pdf
+Texas: Data Privacy and Security Act,05282023,State level sectoral,Retail & E-Commerce,Texas,./pdfs/Texas/TDPSA.pdf,TDPSA.pdf
+Texas: Act relating to maternal mortality information and data registry,09012025,Comprehensive State level,Health,Texas,./pdfs/Texas/HB 1989 Maternal Mortality.pdf,HB 1989 Maternal Mortality.pdf
