--- data_privacy_law/llm_manager/data/bill_info.csv
@@ -1,25 +1,0 @@
-Title,Date,Type,Sector,State,Path,Filename
-Texas: Act prohibiting social media use by children,09012025,State level sectoral,Children’s Data Protection,Texas,./pdfs/Texas/HB 186 Social_media_children.pdf,HB 186 Social_media_children.pdf
-Texas: Act prohibiting search warrants for menstrual health data,09012025,State level sectoral,Health,Texas,./pdfs/Texas/HB 1426 Mentrual Health Info.pdf,HB 1426 Mentrual Health Info.pdf
-Texas: Act relating to maternal mortality information,09012025,Comprehensive State level,Health,Texas,./pdfs/Texas/HB 1989 Maternal Mortality.pdf,HB 1989 Maternal Mortality.pdf
-Texas: Act relating to patient medical records and discharge duties,12012025,State level sectoral,Health,Texas,./pdfs/Texas/HB 2264 Medical Records.pdf,HB 2264 Medical Records.pdf
-Texas: Act relating to reproductive health information collection and civil penalties,09012025,State level sectoral,Health,Texas,./pdfs/Texas/HB 2831 Reproductive health info.pdf,HB 2831 Reproductive health info.pdf
-Texas: Act relating to smart device data collection transparency,09012025,State level sectoral,Telecommunications & Technology,Texas,./pdfs/Texas/SB 726 Smart Devices.pdf,SB 726 Smart Devices.pdf
-Texas: Data Privacy and Security Act,05282023,State level sectoral,Retail & E-Commerce,Texas,./pdfs/Texas/TDPSA.pdf,TDPSA.pdf
-Washington: Protecting Washington children online,02072025,Comprehensive State level,,Washington,./pdfs/Washington/SB 5708 children.pdf,SB 5708 children.pdf
-Washington: Act Relating to Protecting Washington Children Online,01012026,State level sectoral,Children’s Data Protection,Washington,./pdfs/Washington/HB 1834 Children.pdf,HB 1834 Children.pdf
-Washington: Law on Employer Use of Technology in Workplace,01282025,State level sectoral,Employment & HR,Washington,./pdfs/Washington/HB 1672 Employers data collection.pdf,HB 1672 Employers data collection.pdf
-Washington: Personal Data Privacy Act,01282025,Comprehensive State level,,Washington,./pdfs/Washington/HB 1671 Personal Data Privacy.pdf,HB 1671 Personal Data Privacy.pdf
-Washington: Act Relating to Patients' Health Care Information,01142025,State level sectoral,Health,Washington,./pdfs/Washington/SB5254 Patient Health Data.pdf,SB5254 Patient Health Data.pdf
-Washington: Data Broker Registry and Severance Tax,02072025,State level sectoral,Telecommunications & Technology,Washington,./pdfs/Washington/Hb 1887 data broker registry.pdf,Hb 1887 data broker registry.pdf
-California Privacy Rights Act of 2020: Amending Civil Code,01012023,Comprehensive State level,,California,./pdfs/Comprehensive/CaliforniaPrivacyRightsAct2020.pdf,CaliforniaPrivacyRightsAct2020.pdf
-Oregon: Consumer Personal Data Protection Act,06222023,State level sectoral,Children’s Data Protection,Oregon,./pdfs/Comprehensive/OregonConsumerPrivacyAct.pdf,OregonConsumerPrivacyAct.pdf
-Tennessee Information Protection Act: Amending Tennessee Code Annotated,00000000,Comprehensive State level,,Tennessee,./pdfs/Comprehensive/TenesseInformationProtectionAct.pdf,TenesseInformationProtectionAct.pdf
-Federal level: Comprehensive State level legislation,,Comprehensive State level,,,./pdfs/Comprehensive/TexasDataPrivacy.pdf,TexasDataPrivacy.pdf
-Iowa: An Act Relating to Consumer Data Protection,03282023,State level sectoral,Children’s Data Protection,Iowa,./pdfs/Comprehensive/IowaConsumerDataProtectionAct.pdf,IowaConsumerDataProtectionAct.pdf
-Montana: Establishing the Consumer Data Privacy Act,04152023,State level sectoral,Children’s Data Protection,Montana,./pdfs/Comprehensive/MontanaConsumerDataProtectionAct.pdf,MontanaConsumerDataProtectionAct.pdf
-Virginia: Consumer Data Protection Act,01012025,Comprehensive State level,,Virginia,./pdfs/Comprehensive/VirginiaDataProtectionAct2023.pdf,VirginiaDataProtectionAct2023.pdf
-California: Consumer Privacy Act of 2018,01012025,Comprehensive State level,,California,./pdfs/Comprehensive/California2018.pdf,California2018.pdf
-Connecticut: Act Concerning Personal Data Privacy and Online Monitoring,05102022,Comprehensive State level,,Connecticut,./pdfs/Comprehensive/ConnecticutPersonalDataPrivacyandOnlineMonitoring.pdf,ConnecticutPersonalDataPrivacyandOnlineMonitoring.pdf
-Colorado: Concerning Additional Protection of Data Relating to Personal Privacy,07012023,Comprehensive State level,,Colorado,./pdfs/Comprehensive/ColoradoPrivacyAct.pdf,ColoradoPrivacyAct.pdf
-Utah: Consumer Privacy Act of 2022,12312023,State level sectoral,Retail & E-Commerce,Utah,./pdfs/Comprehensive/UtahConsumerPrivacyAct.pdf,UtahConsumerPrivacyAct.pdf
